+++ ricaun.NUnit/Services/TestService.cs
@@ -22,7 +22,7 @@ namespace ricaun.NUnit.Services
         /// </summary>
         public Type[] IgnoreAttributes { get; set; } = new[] {
             typeof(IgnoreAttribute),
-            typeof(TestCaseAttribute),
+            //typeof(TestCaseAttribute),
             typeof(ExplicitAttribute)
         };
 
@@ -51,17 +51,6 @@ namespace ricaun.NUnit.Services
         #region public
 
         /// <summary>
-        /// GetTestMethods
-        /// </summary>
-        /// <param name="type"></param>
-        /// <returns></returns>
-        private IEnumerable<MethodInfo> GetTestMethods(Type type)
-        {
-            var methods = type.GetMethods().OrderBy(e => e.Name);
-            return methods.Where(AnyTestAttribute).Where(e => TestEngineFilter.HasName(e.Name));
-        }
-
-        /// <summary>
         /// Test
         /// </summary>
         /// <returns></returns>
@@ -80,17 +69,7 @@ namespace ricaun.NUnit.Services
                     return testType;
                 }
 
-                var testMethods = methods.Where(AnyAttributeName<TestAttribute>);
-
-                //if (onlyReadTest)
-                //{
-                //    foreach (var testMethod in testMethods)
-                //    {
-                //        var testModel = new TestModel() { Name = testMethod.Name };
-                //        testType.Tests.Add(testModel);
-                //    }
-                //    return testType;
-                //}
+                var testMethods = methods.Where(AnyTestAttribute);
 
                 var methodOneTimeSetUp = methods.FirstOrDefault(AnyAttributeName<OneTimeSetUpAttribute>);
                 var methodOneTimeTearDown = methods.FirstOrDefault(AnyAttributeName<OneTimeTearDownAttribute>);
@@ -103,8 +82,11 @@ namespace ricaun.NUnit.Services
                 {
                     foreach (var testMethod in testMethods)
                     {
-                        var testModel = InvokeTestInstance(testMethod, methodSetUp, methodTearDown);
-                        testType.Tests.Add(testModel);
+                        foreach (var nUnitAttribute in GetMethodTestAttributes(testMethod))
+                        {
+                            var testModel = InvokeTestInstance(testMethod, methodSetUp, methodTearDown, nUnitAttribute);
+                            testType.Tests.Add(testModel);
+                        }
                     }
                 }
                 var downOneResult = InvokeResultInstance(methodOneTimeTearDown);
@@ -183,6 +165,49 @@ namespace ricaun.NUnit.Services
             return test;
         }
 
+        private TestModel InvokeTestInstance(MethodInfo method, MethodInfo methodSetUp, MethodInfo methodTearDown, NUnitAttribute nUnitAttribute)
+        {
+            var test = new TestModel();
+            test.Name = method.Name;
+            test.Alias = GetMethodTestName(method, nUnitAttribute);
+            test.Success = true;
+
+            using (var console = new ConsoleWriterDateTime())
+            {
+                if (IgnoreTest(method, out string messageIgnore))
+                {
+                    test.Message = messageIgnore;
+                    test.Console = console.GetString();
+                    test.Time = console.GetMillis();
+                    test.Skipped = true;
+                    return test;
+                }
+
+                var upResult = InvokeResultInstance(methodSetUp);
+                var methodResult = upResult;
+                if (upResult.Success)
+                {
+                    if (nUnitAttribute is TestCaseAttribute testCaseAttribute)
+                        methodResult = InvokeResultInstanceTestCase(method, testCaseAttribute);
+                    else
+                        methodResult = InvokeResultInstance(method);
+                }
+                var downResult = InvokeResultInstance(methodTearDown);
+
+                var success = upResult.Success & methodResult.Success & downResult.Success;
+                var message = string.Join(Environment.NewLine, upResult.Message, methodResult.Message, downResult.Message).Trim();
+
+                test.Success = success;
+                test.Skipped = methodResult.Skipped;
+                test.Message = message;
+                test.Console = console.GetString();
+                test.Time = console.GetMillis();
+            }
+
+            return test;
+        }
+
+
         #endregion
 
         #region InvokeResult
@@ -253,6 +278,66 @@ namespace ricaun.NUnit.Services
             var equals = (value is not null) ? value.Equals(expectedResult) : expectedResult is null;
             return equals;
         }
+
+        private bool IsValueExpectedResult(TestCaseAttribute testCase, object value, out object expectedResult)
+        {
+            expectedResult = null;
+            if (testCase is null)
+                return true;
+
+            expectedResult = testCase.ExpectedResult ?? null;
+            var equals = (value is not null) ? value.Equals(expectedResult) : expectedResult is null;
+            return equals;
+        }
+
+        private InvokeResult InvokeResultInstanceTestCase(MethodInfo method, TestCaseAttribute testCase)
+        {
+            var result = new InvokeResult();
+            if (method is null)
+                return result;
+
+            if (IgnoreTest(method, out string message))
+            {
+                result.Message = message;
+                result.Skipped = true;
+                return result;
+            }
+
+            try
+            {
+                var value = Invoke(this.instance, method, testCase.Arguments);
+
+                if (IsValueExpectedResult(testCase, value, out object expectedResult) == false)
+                {
+                    result.Success = false;
+                    result.Message = $"Expected:\t{expectedResult}\tBut was:\t{value}";
+                }
+
+            }
+            catch (Exception ex)
+            {
+                var exInner = ex.InnerException is null ? ex : ex.InnerException;
+                result.Success = false;
+                result.Message = exInner.ToString();
+                switch (exInner)
+                {
+                    case SuccessException success:
+                        result.Success = true;
+                        result.Message = success.Message;
+                        break;
+                    case IgnoreException ignore:
+                        result.Success = true;
+                        result.Skipped = true;
+                        result.Message = ignore.Message;
+                        break;
+                    case AssertionException assertion:
+                        result.Message = assertion.Message + Environment.NewLine + assertion.ToString();
+                        break;
+                }
+            }
+            return result;
+        }
+
         #endregion
     }
 }