+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.0.7] / 2022-11-30
+## [1.0.7] / 2022-12-01
 ### Features
+- Add `TestCaseAttribute` suport
+- Add `Alias` in `TestModel`
 - Add `ExplicitEnabled` in Filter
 
 ## [1.0.6] / 2022-11-30 - 2022-10-31